TRACKR-297 Fixes GMail Configuration
Spring Boot 1.3.3 introduced a circular dependency of beans in trackr.
The Spring Mail autoconfiguration autowired a session and exposed a JavaMailSender
while our configuration autowired the JavaMailSender and exposed the session.

We need the session for the Spring Integration configuration as a bean. The solution
is to create the JavaMailSender ourselves. This is done by copying the relevant parts
out of the Spring auto configuration and disabling it.